// src/model/category.model.ts
export class CategoryModel {
  changed: any[] = [];
  default_langcode: any[] = [];
  description: any[] = [];
  langcode: any[] = [];
  name: any[] = [];
  parent: any[] = [];
  path: any[] = [];
  revision_created: any[] = [];
  revision_id: any[] = [];
  revision_log_message: any[] = [];
  revision_translation_affected: any[] = [];
  revision_user: any[] = [];
  status: any[] = [];
  tid: any[] = [];
  uuid: any[] = [];
  vid: any[] = [];
  weight: any[] = [];
}